+++ src/ponos_load_generator.erl
@@ -58,11 +58,11 @@
           limit_reported    :: boolean(),
           name              :: ponos:name(),
           next_trigger_time :: number(),
-          intensities       :: list(erlang:now()),
-          intensity         :: ponos:intensity(),
+          intensities       :: list(erlang:timestamp()),
+          intensity         :: ponos:intensity() | undefined,
           running_tasks     :: integer(),
           start             :: erlang:timestamp(),
-          task_runner       :: module(),
+          task_runner       :: {module(), any()},
           task              :: ponos:task(),
           tick_counter      :: integer()
          }).
@@ -317,7 +317,7 @@ do_prune_intensities(State) ->
       Fun = fun(E) ->
                 time_passed_in_ms(Now, E) < ?PRUNE_INTENSITY_INTERVAL
             end,
-      NewIntensities = [I1, I2 
+      NewIntensities = [I1, I2
                         | lists:takewhile(Fun, T)],
       state_set_intensities(State, NewIntensities);
     _LessThanTwoElements ->
@@ -430,7 +430,7 @@ calc_current_load(State) ->
       0.0
   end.
 
-do_calc_current_load(_State, 0)     -> 0.0;
+do_calc_current_load(_State, 0.0)   -> 0.0;
 do_calc_current_load(State, Period) ->
   Intensities = state_get_intensities(State),
   _CurrentLoad = min(state_get_intensity(State),(length(Intensities) - 1) / Period * 1000.0).
@@ -438,12 +438,7 @@ do_calc_current_load(State, Period) ->
 calc_top_modeled_load(State) ->
   Start = state_get_start(State),
   LoadSpec = state_get_load_spec(State),
-  case Start of
-    undefined ->
-      0.0;
-    Start ->
-      LoadSpec(trunc(time_passed_in_ms(os:timestamp(), Start)))
-  end.
+  LoadSpec(trunc(time_passed_in_ms(os:timestamp(), Start))).
 
 time_passed_in_ms(Now, Then) ->
   timer:now_diff(Now, Then) / 1000.
@@ -506,11 +501,12 @@ state_clear_limit_reported(State) ->
 
 %%%_* EUnit Tests ======================================================
 -ifdef(TEST).
+-ifndef(DIALYZER_RUN).
 
 calc_current_load_test_() ->
   [ ?_assertEqual(0.0, calc_current_load(#state{intensities=[]}))
-  , ?_assertEqual(0.1, do_calc_current_load(#state{intensities=[1,2]}, 10000))
-  , ?_assertEqual(0.0, do_calc_current_load(#state{intensities=[1]}, 0))
+  , ?_assertEqual(0.1, do_calc_current_load(#state{intensities=[1,2]}, 10000.0))
+  , ?_assertEqual(0.0, do_calc_current_load(#state{intensities=[1]}, 0.0))
   ].
 
 duration_is_exceeded_test_() ->
@@ -622,6 +618,7 @@ mk_state(Intensity, TriggerTime) ->
 mk_state(Intensity, TriggerTime, LoadSpec) ->
   (mk_state(Intensity, TriggerTime))#state{load_spec = LoadSpec}.
 
+-endif.
 -endif.
 
 %%%_* Emacs ============================================================